requirements.txt
huggingface-hub==0.29.3
ultralytics==8.3.93
supervision==0.25.1
pillow==11.1.0
